+++ data/src/appearance/theme.rs
@@ -265,6 +265,7 @@ pub struct Text {
     pub info: OptionalTextStyle,
     pub debug: OptionalTextStyle,
     pub trace: OptionalTextStyle,
+    pub spellcheck_misspelled: OptionalTextStyle,
 }
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, Default)]
@@ -847,6 +848,7 @@ mod binary {
         BufferServerMessagesRequestTopic = 57,
         ButtonsPrimaryBorderActive = 58,
         ButtonsSecondaryBorderActive = 59,
+        TextSpellcheckMisspelled = 60,
     }
 
     impl Tag {
@@ -866,6 +868,9 @@ mod binary {
                 Tag::TextError => styles.text.error.color,
                 Tag::TextWarning => styles.text.warning.color?,
                 Tag::TextInfo => styles.text.info.color?,
+                Tag::TextSpellcheckMisspelled => {
+                    styles.text.spellcheck_misspelled.color?
+                }
                 Tag::TextDebug => styles.text.debug.color?,
                 Tag::TextTrace => styles.text.trace.color?,
                 Tag::BufferAction => styles.buffer.action.color,
@@ -1001,6 +1006,9 @@ mod binary {
                 Tag::TextError => styles.text.error.color = color,
                 Tag::TextWarning => styles.text.warning.color = Some(color),
                 Tag::TextInfo => styles.text.info.color = Some(color),
+                Tag::TextSpellcheckMisspelled => {
+                    styles.text.spellcheck_misspelled.color = Some(color);
+                }
                 Tag::TextDebug => styles.text.debug.color = Some(color),
                 Tag::TextTrace => styles.text.trace.color = Some(color),
                 Tag::BufferAction => styles.buffer.action.color = color,