+++ src/buffer.rs
@@ -796,7 +796,9 @@ impl Buffer {
     pub fn insert_user_to_input(
         &mut self,
         nick: Nick,
+        clients: &data::client::Map,
         history: &mut history::Manager,
+        config: &Config,
         autocomplete: &Autocomplete,
     ) {
         match self {
@@ -810,18 +812,24 @@ impl Buffer {
             Buffer::Server(state) => state.input_view.insert_user(
                 nick,
                 state.buffer.clone(),
+                clients,
+                config,
                 history,
                 autocomplete,
             ),
             Buffer::Channel(state) => state.input_view.insert_user(
                 nick,
                 state.buffer.clone(),
+                clients,
+                config,
                 history,
                 autocomplete,
             ),
             Buffer::Query(state) => state.input_view.insert_user(
                 nick,
                 state.buffer.clone(),
+                clients,
+                config,
                 history,
                 autocomplete,
             ),
@@ -1303,6 +1311,7 @@ impl Buffer {
 
     pub fn clear_draft_reply(
         &mut self,
+        clients: &data::client::Map,
         history: &mut history::Manager,
         config: &Config,
     ) -> bool {
@@ -1316,16 +1325,19 @@ impl Buffer {
             | Buffer::ConfigEditor(_) => false,
             Buffer::Server(state) => state.input_view.clear_draft_reply(
                 &state.buffer,
+                clients,
                 history,
                 config,
             ),
             Buffer::Channel(state) => state.input_view.clear_draft_reply(
                 &state.buffer,
+                clients,
                 history,
                 config,
             ),
             Buffer::Query(state) => state.input_view.clear_draft_reply(
                 &state.buffer,
+                clients,
                 history,
                 config,
             ),