+++ src/screen/dashboard.rs
@@ -1531,7 +1531,7 @@ impl Dashboard {
                         // close_picker does not return true.
                         } else if state
                             .buffer
-                            .clear_draft_reply(history, config)
+                            .clear_draft_reply(clients, history, config)
                         {
                             return (Task::none(), None);
                         }
@@ -2229,7 +2229,9 @@ impl Dashboard {
                         {
                             pane.buffer.insert_user_to_input(
                                 nick,
+                                clients,
                                 history,
+                                config,
                                 &config.buffer.text_input.autocomplete,
                             );
                         }